+++ diesel_compile_tests/tests/fail/boxed_clone_queries_require_selectable_expression_for_order.rs
@@ -1,0 +1,25 @@
+extern crate diesel;
+
+use diesel::pg::Pg;
+use diesel::*;
+
+table! {
+    users {
+        id -> Integer,
+        name -> VarChar,
+    }
+}
+
+table! {
+    posts {
+        id -> Integer,
+        title -> VarChar,
+    }
+}
+
+allow_tables_to_appear_in_same_query!(users, posts);
+
+fn main() {
+    users::table.into_boxed_clone::<Pg>().order(posts::title.desc());
+    //~^ ERROR: type mismatch resolving `<table as AppearsInFromClause<table>>::Count == Once`
+}